layers:Only flag one aspect error at a time
For image aspect bit errors, first flag the error if correct bit is not
set, else flag error if additional bits are incorrectly set. There's
not a real benefit to flagging both errors and it complicates testing.
Generally the first error should be sufficient to allow a user to
correctly update their aspect bit usage.